Makefile: Include libwpa_client.so only with hostapd-utils package
With qcamain premium profile, we are the seeing the below compilation
issue. This is due to hostapd-utils is not enabled in .config as it
is related to open profile. So, include the libwpa-client.so only
when the hostapd-utils package is enabled in the .config.

Compilation error:
cp: cannot stat `/home/build_server/qcamain_premium/qsdk/build_dir/
target-arm_cortex-a7_uClibc-1.0.14_eabi/hostapd-supplicant-full/
hostapd-2019-11-04/wpa_supplicant/libwpa_client.so': No such file or directory

Change-Id: If1f5ed44f48f254337a5fddac3a6ea40396063f9
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>