hostapd: Multiple BSSID DTIM setting
For EMA (Enhanced Multiple BSSID Advertisement) AP, nontransmitted
profiles should have DTIM value as a multiple of profile periodicity
as mentioned in IEEE P802.11ax/D7.0, 11.1.3.8.3 Discovery of a
nontransmitted BSSID profile.

This is to make sure that if there is a change in a profile, the next
beacon containing that profile will be the DTIM beacon.
The stations associated with the profile will be notified of the change
as soon as possible.

With existing design, the transmitted beacon is set before RSN
information element is formed for any nontransmitted profile hence the
beacon has these profiles with open encryption.
It also sets wrong profile periodicity until all profiles are up.

This patch fixes the issue by retrieving configurations for all
nontransmitted profiles before setting beacon for the transmitted
profile.
The profile periodicity is calculated before setting the beacon
and DTIM values are set to be the next multiple of profile periodicity.

Signed-off-by: Aloka Dixit <alokad@codeaurora.org>
Change-Id: I04ef8e24f137a4a71afabd73949d4865a495e158