hostapd: Fix lowest association ID with multiple BSSID element
Association ID space is shared when multiple BSSID element is included
in beacons.

As per IEEE P802.11ax/D6.0, 11.1.3.8.5 Traffic advertisement in a
multiple BSSID set, the lowest association ID to be assigned to non-AP
station should be 2^n, where n is the maximum BSSID indicator of the
multiple BSSID set.
Example, if the set has 6 BSSes, the lowest association ID will be 8.

This patch sets the correct value which was incorrectly set to the
number BSSes in the multiple BSSID set.

Signed-off-by: Aloka Dixit <alokad@codeaurora.org>
Change-Id: I2b86d0cc0bc02cb2db8168281d74a8e371ee9367