hostapd: Hidden SSID support in multiple BSSID IE
Hidden nontransmitted BSSID profiles will be included in the beacons and
probe responses but SSID value will be removed or set to all zeros
depending on the configured value of 'ignore_broadcast_ssid'.
If complete profiles are omited, clients cannot stay connected to the AP.

For unicast probe requests with SSID set to a hidden nontransmitted BSSID
profile, complete SSID is included in the probe response.
This patch adds a new input parameter in related functions, pointer to
hostapd_data for nontransmitted profiles, to differentiate this case.

Signed-off-by: Aloka Dixit <alokad@codeaurora.org>
Change-Id: I69123f0d5ef899f9c6546f3f544a7a509327bca2