hostapd: always respawn hostapd instance immediately
The default value for respawn is 3600 5 5 which means that if
hostapd is killed 5 times within 5 mins, the service stopped.
This breaks some automation when the script kills hostapd instance
so often for each iteration. Setting the retry value to 0 will cause
procd to try restarting the service indefinitely and setting
timeout to 1 so that instance will be restarted immediately.

Change-Id: Ifb029470ba6e678dbca8b90756ccf13f8afca983
Signed-off-by: Rajkumar Manoharan <rmanohar@codeaurora.org>