hostapd: correct 20Mhz PSD calculation
20MHz PSD value should be set to 2 * (default tx-power value of -1
dBm), which results to signed value of 0xfe. Since PSD value is added
to RNR IE as an 8 bit unsigned value but is interpreted as a 8 bit 2s
complement value, the default tx-power value is stored as unsigned
value of 255. To set 20MHz PSD value, the unsigned default value is
subtracted by one to produce the signed value of 0xfe.

Signed-off-by: Muna Sinada <msinada@codeaurora.org>
Change-Id: I4c80fca7dea42895109662d2aa79ce6097abd83a