hostapd: Makefile change to skip the MD5SUM check for hostapd tarball
We don't need to check the md5sum/hash values for every compilation,
which leads to download the hostapd tarball from upstream hostapd.git
even we already have the hostapd tarball in the dl directory.

So, we can use the “skip” value for md5sum/hash during compilation
which solves the above issue.

Change-Id: I2978f9bfcd1745d6237cc1b4b2c85345d639288c
Signed-off-by: Maharaja Kennadyrajan <mkenna@codeaurora.org>