hostapd: disabling auto scan for the mesh mode in wpa_supplicant
noscan disables the scanning of bandwidth availablity check
for 40Mhz in hostapd. When in mesh mode with encryp wpa_supplicant
is used, here 'disable_40mhz_scan' is the term used for noscan in
wpa_supplicant.

So setting 'disable_40mhz_scan' in wpa supp conf when noscan is enabled
in wireless config.

Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>
Change-Id: Ia6204eaeb2cca3874c5dd01f5ce589dd9fba3cb7