hostapd: add RNR by default in FILS discovery for 6GHz-only mode
This commit modifies is_6ghz_colocated() to return 6GHz modes (no 6GHz,
6GHz-only, 6GHz coloated with lower band and lower band colocated with
6GHz) instead of true/false.
This is added so that in 6GHz only mode, reduced neighbor element will
be
included in FILS discovery by default if there are more than one BSS on
the same radio.
This commit also updates 6GHz beacons when 2.4/5GHz interfaces are
brought up so that RNR will be removed from FILS discovery frames as
the APs are not running in 6GHz-only mode anymore.
The code flow is changed from if-else to a switch block in
hostapd_eid_rnr_len() and hostapd_eid_rnr() to handle different types of
frames better.

Signed-off-by: Aloka Dixit <alokad@codeaurora.org>
Change-Id: I6d578503ce87dd459206916453d953b77391dc22