hostapd: Fix HE chan switch command to use proper BW
VHT is not enabled when HE mode is indicated in channel
switch command, due to which VHT operation IE is missing
in beacon causing STA to connect in HT40 mode.

HE support for channel switch command patch is modified in
upstream commit causing this issue. It is reverted to the
previous version for fixing the same.

Fixes: I4bd28058adbc27 (hostap: upgrade hostapd to 59e9794c7d7e)

Signed-off-by: Lavanya Suresh <lavaks@codeaurora.org>
Change-Id: Idf916eab581cd99b911acdde8ac6c05aa5cde4d6