hostapd: Mesh mode 6ghz radio unable to set the channel 209 in HE80
In mesh mode 6ghz radio, channel 209 with HE80 is not setting properly instead it
is choosing to operate in 20Mhz NO_HT mode.

This is because while checking for max channel bw availability in bw160
if the channel is not avail to operate in 160Mhz it is choosing to
operate in NO_HT though the same channel is available to operate in 80Mhz.

Fixed by checking max channel bw availablity for bw160 only if this
channel is available in bw160 list.

Change-Id: Iba1900f371c1d3e417419c4e999cb3c9c4d30923
Signed-off-by: Karthikeyan Kathirvel <kathirve@codeaurora.org>