hostapd: add support for 6GHz operation
6 GHz gives users ability to select the AP mode from LPI, SP and
VLP modes. This mode needs to be passed on to the cfg80211
for further processing during AP bring up. Also, the regulatory
rules received by hostapd contains psd values for 6G rules.
Hostapd needs to store these values in order to later advertise
it in the tx power element in the beacon.

This patch adds the support to send the user configured 6g power
type during set_channel command and as well to get and store the
psd values from reg rules which will be later on used in tpe
advertisement during beacon formation.

Change-Id: I5ae0e968004b762d40e4823a0876192f3fb8df42
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>