hostapd : fix random channel selection in awgn
The available channel list in hostapd holds the current operating channel also.

So sometime while choosing a random channel, the existing logic chooses current
operating channel as new random channel.

And since the new channel is same as current operating channel,
channel switch fails in hostapd and AP remains in current operating channel
instead of moving to a new random channel.

Support is added to ignore the current operating channel to be chosen
as new random channel.

And when chan_bw_interference_bitmap is 0, the host is expected to
do channel change. support is added for the same.

Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>
Change-Id: Ie1b18ef308cad04883c6bda1dfa1422ee7ca8b9b