hostapd: fix missing functionalities in RU puncturing
Fix following two issues related to RU puncturing:

- Legacy modes (VHT, HE) should advertise downgraded bandwidth if
RU puncturing is enabled in EHT mode. This is required for the legacy
clients which cannot parse the EHT operation elements hence do not
support EHT RU puncturing.

- Beacons after channel switch command didn't include the new RU
puncturing bitmap as this data was mising from the event.
Modify structures and functions to add the missing data in beacons.

Additionally, move the validation of RU puncturing bitmap from
q02-043-acs-validate-the-RU-puncturing-bitmap.patch to
q02-034-ru_puncturing-add-bitmap-to-EHT-operation-element.patch and
exercise this during initial AP bring up as well as channel switch
operation. No change in the ACS functionality.

Change-Id: I5f5fb4c74db9ad4d7d4e68188c8a2b2947a48a7c
Signed-off-by: Aloka Dixit <quic_alokad@quicinc.com>