hostapd: Add config to truncate ext capabilities
Certain legacy clients are not able to scan 11ax vaps due to
extended capabilities which are more than 8bytes in length.
Hence added a work around to trucate the ext caps to 8bytes
based on the hostapd config ext_cap_len.
Legacy clients are able to scan and connect if hostapd config
as ext_cap_len=8, ieee80211ac=1, ieee80211ax=0

Change-Id: Ibd17f300b6ef88e4c716e703dbc2ba63d1f1cef5
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>