hostapd: Add config to truncate ext capabilities
Certain legacy clients are not able to scan 11ax vaps due to
extended capabilities which are more than 8 bytes in length.
Hence added a work around to trucate the ext caps to 8 bytes
based on the hostapd config ext_cap_len.
Legacy clients are able to scan and connect if hostapd config
as ext_cap_len=8, ieee80211ac=1, ieee80211ax=0

Change-Id: Iddfb906be547aae46f1b7a65a2a1084ae02f3535
Signed-off-by: Sowmiya Sree Elavalagan <quic_ssreeela@quicinc.com>
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>