hostapd: Fix incorrect bandwidth in EHT Operation IE for AP+STA mode
Currently EHT operation IE has bandwidth set to 40 MHz when AP
is actually on 20 MHz.
I58a789a1d8e("hostapd: Fix EHT Association in 20MHz") tries to
distinguish 20 MHz bandwidth to 40 MHz by seg0 field and this
causes incorrect bandwidth when seg0 is not set. Fix this by
determining operation bandwidth from op_class in 6 GHz band
and secondary_channel in 5 GHz and 2 GHz bands.
secondary_channel is determined from HT capabilities which is
not present in 6 GHz band and op_class in not mandatory to be
set in 5 GHz and 2 GHz bands.

Change-Id: I46e99c5f94eb6dec1cce2dc889d79a536d6425b3
Signed-off-by: Pradeep Kumar Chitrapu <quic_pradeepc@quicinc.com>