hostapd : fix he40/eht40 bringup with acs
when ap is brought up in he40/eht40 with acs,
the ap comes up with 20 Mhz bandwidth. It is expected to
come up with 40 Mhz bandwidth.

conf->secondary_channel doesnot hold the correct value and it leads
to choosing 20 Mhz in the api hostapd_set_freq_params.

when the config acs is chosen, the hostapd config
he_oper_centr_freq_seg0_idx is not filled with correct value
as the channel is not known during bring up.

So using the config he_oper_centr_freq_seg0_idx to fill the
conf->secondary_channel doesnot work with acs.

So use op_class to determine the bandwidth and based
on the bandwidth fill the conf->secondary_channel.

Change-Id: Ife739b229080f434ca755ac4bb047652ae11b03c
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>