hostapd: Add frequency info in start ap
In case of single hw multiple band scenario, when we
bring up multiple virtual interface in various bands
simultaneously then the driver maps a random channel
to the virtual interface when start ap command is sent
as freq info is not present.

Hence, add the frequency information in start ap netlink
command so that the driver maps the appropriate channel
context by parsing it.

Change-Id: Ie3dcd6815d436c8baf9f91e917ba378f67d50cc9
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>