hostapd: Fix nl80211 enum to align with ath12k
Recent TPC changes, added the flag to set PSD to
channel flag. NL80211 attribute to advertise the
PSD capability in ath12k is not matching the enum
in hostapd which caused hostapd to think
NL80211_FREQUENCY_ATTR_NO_320MHZ is set and hence
was not allowing APUT to come up in 320MHz.

Fixed by re ordering NL80211 attribute in hostapd to
match with ath12k

Change-Id: Id1226acc22cb685b95c66c7455221151e5199eb5
Signed-off-by: Ramya Gnanasekar <quic_rgnanase@quicinc.com>