hostapd: add MLD Single link operation support
As MLD operation for a AP is mandatory as per
ieee80211-be std, include support for handling
single link operation as part of Multi link support

This is basically a stripped off version of a
MLO support with support and infrastructure for
handling ML IEs which are minimally required for
a AP to operate as a MLD and support single link
operation. No link info is added or parsed and
such assoc reqs if received are rejected.

To keep the changes minimal, the sta object uses
mld address as its address and link address is
additionally included for any operation required
specifically, but assumption is hostapd and cfg80211.
mac80211 will operate using mld address and assumes
driver performs the link<->mld address conversion
properly.

Change-Id: I09f338cd37ce77055c6b6685fa08a0cdd58315e6
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>
Signed-off-by: Dinesh Karthikeyan <quic_dinek@quicinc.com>