hostapd: [single wiphy] avoid channel switch across radio
In single wiphy scenario, the phy supports more than one band
and when we try to change channel without specifying the wifi
interface in hostapd_cli, it would lead to an exception.

To handle this, if the given freq is not in the list of
channels for that interface's current operating mode,
then the cli would fail.

Hence, in case of single wiphy, the below hostapd_cli should be
used to change the channel

hostapd_cli -i wlanX chan_switch <cs count> <freq> <caps> \
center_freq1=<value> bandwidth=<bw>

Change-Id: Ibf03cc0d7b8b32592f64d96d3c44381ccf545e09
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>