hostapd : awgn - fix new channel selection logic
From the list of channels allowed by Australia,
only one 320 MHz operable channel is available.

So if awgn is detected when operating in 320 MHz, the ap
needs to reduce its bandwidth to next max available bw 160 MHz.

But with current design if no other 320 Mhz channel is available,
hostapd bails out with a log saying that no channel with 320 MHz is available.
we don't check for next max available bandwidth in our current design and
this leads to ap to remain in the same channel even if awgn is detected.

So the channel selection logic is changed in a way that the next max
available bandwidth is searched if no new channel is available for current
bandwidth. The search continues till 20 MHz bandwidth.

This generic logic will help to fix similar issues in future.

Also optimized the channel selection logic by looping over the list of all
the channels only once.

Change-Id: I07fdde29ca0ee20a1fbe14d2023a494a38c67a5b
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>