Changes to fix compiltaion error for 32-bit architecture
Depending on the type of architecture sigcontext-32.h will
be included for 32 bit builds and sigcontext-64.h in 64
bit builds because of the absence of linux/types.h in
sigcontext-32.h we will face the error __s8 undfined.So
the declaration for this var is added in necessary source
files.

Tested both 32 and 64 bit builds also tested AP bringup and
sanity.

Change-Id: I9262f5f1747ead47067ad0e9ea9273564db96b87
Signed-off-by: Avula Sri Charan <quic_asrichar@quicinc.com>