hostapd: DFS interference bitmap support during DFS interference
Currently when DFS interference is detected the whole configured
frequency is moved to unavailable(NOL) state.

These changes adds support to parse a new NL attribute
"NL80211_ATTR_RADAR_BITMAP" which gets the bitmap.This bitmap
indicates the sub-channels where DFS interference is detected.
In the bitmap, the LSB denotes the lowest 20MHz based on the
center frequency configured. With this bitmap, the affected
sub-channels alone will be moved to unavailable state.

Change-Id: I5ca923b00dd594485a0bed354c96c4754ef1364d
Signed-off-by: Nagarajan Maran <quic_nmaran@quicinc.com>