hostapd: Fix WDS Repeater assoc fail on 6G with 160MHz
While trying to connect the STA with WDS AP, filling the chandef info
in the driver. In the Problematic case, receiving the EHT160 center
frequency is invalid. Due to this chandef validation getting fail and
returns a failure message.

[ 6340.667455] wlan1: authenticate with 00:03:7f:01:58:33
[ 6340.667502] wlan1: HE 6GHz operation resulted in invalid chandef: 6195 MHz/5/0 MHz/0 MHz
[ 6340.671605] wlan1: Rejecting non-HE 6/7 GHz connection
[ 6343.937592] wlan1: authenticate with 00:03:7f:01:58:33

From the analysis, didn't update the EHT160 center frequency information
in hostapd.

Fix it by updating the EHT160 center frequency information.

Change-Id: Ibeb9a1c65de29f13d11bb4a5d59c1db5ac4fedd4
Signed-off-by: Sivashankari Madhavan <quic_sivamadh@quicinc.com>