wpa_supplicant: optimize single wiphy mesh scan
In case of single phy with multiple band scenario,
mesh ping fails for vdev (say 5G) because mesh key
is installed in different band (say 6G).

This is because, when mesh vap is started scan is
triggered without the frequency information.
Hence,on driver side it scans for all bands(say 2G, 5G,
and 6G). For each scan on particular band a vdev is created
with same macaddr on the driver  which leads to incorrect
key installation.

To handle this, add changes to send all the frequencies
corresponding to the band being configured by user

Change-Id: I377d99a4ecadc9b84a0899939ecd39f15695dfb3
Signed-off-by: Harshitha Prem <quic_hprem@quicinc.com>