hostapd: some fixes for cohosted ML bss handling
When multiple MLDs are brought up with configuratios
like 2+5 MLD and a 5+6 MLD, indicating first link for
driver setup based on the head of the MLD might not be
accurate, since the interface would be created when one
of the config is parsed first(where both interface and bsses
will be created) and other link config is parsed, we can
indicate first link or not based on if the ML vif is already
created.

Also, while handling if_add(), the use_existing flag is sent
wrongly as true due to stack garbage. Fix this as well.
This flag can be set even if the MLD is already setup.

Do not reset drv->mlo_link_id on removing first link bss itself,
the link id will be needed for other links as well in case of
multiple bss

Change-Id: I8340b91774b2311be64a289a2409b0fcfdf6386c
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>