hostapd: Add TSF offset placeholder in per sta profile
For a AP MLD, the TSF offset field needs to be present
in the per sta profile indicating the offset between
the reporting and the reported link profile.

Add a placeholder to indicate the TSF offset, currently
set to 0. Later this needs to be filled appropriately
based on the TSF offset between the bsses.

Change-Id: I32795d1b5dde995487ebb18681cf982a6899b786
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>