hostapd: handle DFS detect event during MLO operation
Currently, when DFS interference event is detected, hostapd first
tries to find a non-DFS channel with same bandwidth. If it can not
find one, then it tries to find a DFS channel with same config. If a
channel is found, it will bring down 5 GHz interface and then do a CAC
and post CAC bring up the 5 GHz interface again. If no channel is
available, then it will start reducing the bandwidth and try to see if
DFS_ANY channel is available or not.

This logic will not work in MLO case since we dont have dynamic link
handling support yet.  Hence, we can not just tear down 5 GHz link which
is part of MLO.

To address this -
 * changes are made to first try to reduce bandwidth and try to find a non-
   DFS channel uptil 20 MHz.
 * if no channel is found, then again from the original bandwidth, try to
   find a DFS channel where CAC needs to be done.

Tearing down the MLD interface logic will be added later.

Change-Id: Icf8a07b76677db6ef643f48354580bd80297fea8
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>