hostapd: fix EML and MLD capabilities processing
Currently, hostapd directly tries to parse two NL attributes related
to EML and MLD capabilities respectively. However, these two attributes
are driver related capabilities and might not be advertised for every
driver. In such case if hostapd tries to still parse it, segmentation
fault is seen.

Fix the issue by first validating that the attributes are present and
then only parsing it.

Fixes: f2fd62f70577 ("hostapd: Process EML and MLD capabilites and add to ML IE")

Change-Id: Ic75844fa762424e73d4c00c926156617e8c125c2
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>