hostapd: Fix 2g EHT40 failure
Currently, AP in 2g EHT40, the control field is getting assigned wrong
values. Basically, the control field contains the value 0 instead of 1
because when seg0 is not set, it is not able to distinguish between 20
MHz and 40 MHz.

Hence, determining operation bandwidth from op_class in 6 GHz band and
secondary_channel in 5 GHz and 2 GHz bands.

During package upgrade this change got missed.

Fixes: e50eef7ba2 ("hostapd: Package Upgrade 09/19")

Change-Id: Ie77174072dd8210179b6eb7fcec82c40c34e3252
Signed-off-by: Pradeep Kumar Chitrapu <quic_pradeepc@quicinc.com>
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>
Signed-off-by: Rahul Bhattacharjee <quic_rbhattac@quicinc.com>