hostapd: Fix wpa3-transition security mode association failure
While trying to associate the MLE not supported client with wpa3
security mode, authentication getting fails.

In authentication replay, AP sends the MLE information to the Non-MLE
client so that the client ignores the unknown authentication data reply.

Fix it by avoids to filling in the MLE info for Non-MLE supported clients.

Change-Id: Ib70ab5edf8f0520c47970c5d131177187026308f
Signed-off-by: Sivashankari Madhavan <quic_sivamadh@quicinc.com>