hostapd_cli: fix selection of interface in legacy case
Currently, if -l arguement is not passed in command, it will try to fetch
the interface socket by parsing the socket directory. This leads to selection
of last registered socket.
For example - wlan2, wlan2-1 and wlan2-3 exists and if command is executed
with interface -i wlan2, since -l is not used, it will try to get via parsing
the directory and in that it selects wlan2-3 always which is wrong since in
non-mlo case, user will not provide -l.

Hence, fix this issue by checking if interface name was provided and link_id
was not provided and if interface is non-ml interface then proceed with the
given interface instead of selecting from the directory. If its ML interface
then select as per the exiting logic.

Fixes: defdbf3d93c6 ("hostapd: create link based hapd ctrl sockets")

Change-Id: If596ad8506f80b08a3a5a267169c42a1064e74ea
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>