hostapd: fix multi underlying split hardware channel selection issue
During multi-hardware under singly phy scenario, single channel list is
maintained for a given band which does not clearly tells how many different
underlying hardwares are being used to form that band. For example - in
5 GHz split hardware case, two underlying hardwares 5 GHz Low and 5 GHz High
are being used to enable full 5 GHz band. Now, when channel is selected from
list of channels, it is possible that a channel can be picked which the
hardware does not support (lower chanels can be picked for 5 GHz high
hardware) which is wrong.

However, the kernel does tells us about the underlying hardwares and the
list of supported frequencies.

Hence, add support to get the advertised hw_idx and the corresponding list
of supported frequencies and then decide what is the current hw_idx being
used.

While at it, add logic to check if the selected channel falls under the
current operating underlying hardware  and only if it falls then continues
with the channel.

Change-Id: Ia3faff888383a760bdf1e8b8dc18455a841a5453
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>