hostapd: VAP goes down after injecting invalid dfs interference
Once the puncturing applied for the dfs interference with valid puncturing pattern,
if receive again the dfs interference with invalid then VAP goes down
due to previous valid pattern send to the channel switch.

To fix this when receive invalid dfs interference affected event reset
the exist pattern to zero.

Change-Id: Ie4b5a49cb71f3d186b1ff82ec62fdf72ce8736bd
Signed-off-by: Ramasamy Kaliappan <quic_rkaliapp@quicinc.com>