hostapd: Add Operating Class and Primary channel in FD for non PSC chan
If a non PSC 6GHz channel is configured, duplicate beacons/FD/UBPR
will be transmitted in other 20MHz channels of the current configured
bandwidth to aid in faster scan. In such cases the duplicate FD
needs to carry the Operating Class and Primary channel for clients
to identify the primary non PSC.

As per IEEE802.11 2020,
"The Operating Class subfield specifies the operating class of the Primary Channel of the transmitting AP
(see 9.4.1.36).
The Primary Channel subfield is set to the channel number of the primary channel (see 11.15.2) if the FILS
Discovery frame is transmitted as a non-HT duplicate PPDU; otherwise, the subfield is not present."

Hence, add the Oper class and Primary channel if the current channel
is non PSC.

Change-Id: I5232061017e7051cb6ce2e1365a5cc46a8ca134c
Signed-off-by: Sriram R <quic_srirrama@quicinc.com>