hostapd: fix AWGN detection in 20 MHz bandwidth
Currently the function get_centre_freq_6g() calculates center frequency
for bandwidth 40, 80, 160 and 320 MHz only. Since 20 MHz is left out,
center frequency remains 0 and error is returned to the caller of the
function. Due to this, next channel can not be selected.

Fix this issue by calculating the center frequency in case of 20 MHz
bandwidth as well.

Fixes: 38728fe012d7 ("hostapd : awgn - fix new channel selection logic")

Change-Id: I27d0585f2112afc64d0f8a91eeb7818ba9ceba31
Signed-off-by: Aditya Kumar Singh <quic_adisi@quicinc.com>