hostapd : fix DS parameter in beacon and probe response
Few clients refuse to connect to the 6 GHz AP when the AP
is brought up in non-psc channel.

The clients receive the duplicate beacons of the AP in the psc-channel
and drops the beacon due to the DS parameter available in the
beacon.

From IEEE Std 802.11-2020 9.3.3.2 Beacon frame format,
DSSS Parameter Set:
The element is optionally present.
The DSSS Parameter Set element is present within Beacon frames
generated by STAs using Clause 15, Clause 16, and Clause 18 PHYs.
The element is present within Beacon frames generated by STAs using a
Clause 19 PHY in the 2.4 GHz band.

The mentioned clauses point to 2.4 GHz PHYs.

DSSS : Direct Sequence Spread Spectrum

As the standard mentions that DS parameter is applicable only to 2.4 GHz
band, add a check to include this parameter only to that band in the beacons
and probe response of the AP.

Change-Id: I642dd8332e8fa8db24ec33b61554fdfb4dbf2cd1
Signed-off-by: Hari Chandrakanthan <quic_haric@quicinc.com>