hostapd: Update dynamic punturing bit pattern for radar detected radio alone
In split phy architecture, when radar is detected, AP reduces the
operating bandwidth instead punturing the sub channel in which radar
is detected.

In current scenario, When the radar is detected on a 5 GHz low radio,
nl event is received for both 5 GHz high and low radio.
We check for overlapped channel and process only the 5 GHz low radio event.
But hostapd update the dynamic puncturing bit pattern for both radio 5 GHz
high and low radio.

Fixed to retain the original the ru puncturing  pattern for non radar detected
5 GHz radio.

Change-Id: Ib142dea6ace8ec8e7baeaa27a2b11c47e0b05a76
Signed-off-by: Karthik M <quic_karm@quicinc.com>